fix/docs: set the log level to accept string inputs; update comments inside of the FLHook config
Remove some old charfile encryption code

+++ source/Core/Logger.cpp
@@ -176,36 +176,33 @@ ______ _      _   _             _      _____  _____               _ _
 
 void Logger::Trace(const std::wstring_view str)
 {
-    if (FLHook::GetConfig()->logging.minLogLevel == 0)
+    if (static_cast<int>(FLHook::GetConfig()->logging.minLogLevel) == 0)
     {
         logQueue.enqueue({ LogLevel::Trace, std::wstring(str), _ReturnAddress() });
     }
 }
 void Logger::Debug(const std::wstring_view str)
 {
-    if (FLHook::GetConfig()->logging.minLogLevel <= 1)
+    if (static_cast<int>(FLHook::GetConfig()->logging.minLogLevel) <= 1)
     {
         logQueue.enqueue({ LogLevel::Debug, std::wstring(str), _ReturnAddress() });
     }
 }
 void Logger::Info(const std::wstring_view str)
 {
-    if (FLHook::GetConfig()->logging.minLogLevel <= 2)
+    if (static_cast<int>(FLHook::GetConfig()->logging.minLogLevel) <= 2)
     {
         logQueue.enqueue({ LogLevel::Info, std::wstring(str), _ReturnAddress() });
     }
 }
 void Logger::Warn(const std::wstring_view str)
 {
-    if (FLHook::GetConfig()->logging.minLogLevel <= 3)
+    if (static_cast<int>(FLHook::GetConfig()->logging.minLogLevel) <= 3)
     {
         logQueue.enqueue({ LogLevel::Warn, std::wstring(str), _ReturnAddress() });
     }
 }
-void Logger::Err(const std::wstring_view str)
-{
-        logQueue.enqueue({ LogLevel::Err, std::wstring(str), _ReturnAddress() });
-}
+void Logger::Err(const std::wstring_view str) { logQueue.enqueue({ LogLevel::Err, std::wstring(str), _ReturnAddress() }); }
 
 std::optional<std::wstring> Logger::GetCommand()
 {
